% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_semicolons_in_strings.R
\name{check_for_semicolons_in_strings}
\alias{check_for_semicolons_in_strings}
\title{Handling historical difficulties}
\usage{
check_for_semicolons_in_strings(x)
}
\arguments{
\item{x}{a data frame with a column named \code{log}}
}
\value{
A data frame
}
\description{
Function is called internally by
\code{\link{logstreams_to_data_frame}} and is not exported (i.e. it is not
intended to be called by package users themselves). It checks whether there
are some additional semicolons in the logstreams - if so, some of them don't
mark data fields boundaries. In a case there are such semicolons it converts
the ones that occur between single quota signs (') to colons. This enables to
handle the problem with very early versions of the \emph{logdataLimeSurvey}
applet that did not convert semicolons in the \code{navigator.userAgent} to
colons by itself, what further caused problems while separating data into
columns by \code{\link{logstreams_to_data_frame}}.
}
\seealso{
\code{\link{logstreams_to_data_frame}}
}
