% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cursor_indices.R
\name{compute_cursor_indices}
\alias{compute_cursor_indices}
\title{Computing indicators describing cursor moves}
\usage{
compute_cursor_indices(
  actions,
  respId = any_of(c("id", "token", "respid")),
  screenId = all_of("screen"),
  returnFormat = c("long", "wide"),
  systemInfo = NULL
)
}
\arguments{
\item{actions}{A data frame containing data regarding \emph{actions} -
typically element \emph{actions} of a list returned by
\code{\link{separate_logdata_types}}.}

\item{respId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying respondent.}

\item{screenId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying survey screen.}

\item{returnFormat}{String indicating whether results should be returned in
a \emph{long} (row is respondent-(sub)question) or in a \emph{wide}
(row is a respondent, there is a separate column for each (sub)question)
format. Can be abbreviated.}

\item{systemInfo}{Optionally a data frame containing data regarding
\emph{system information} - typically element \code{systemInfo} of a list
returned by \code{\link{separate_logdata_types}}. \strong{It is required to
compute relative variants of indices.}}
}
\value{
A data frame with columns:
\describe{
  \item{respId}{Column(s) defined by \code{respId}.}
  \item{screenId}{Column(s) defined by \code{screenId}.}
  \item{dX}{Total distance traveled on horizontal axis [px].}
  \item{dY}{Total distance traveled on vertical axis [px].}
  \item{vX}{Average horizontal speed [px/s].}
  \item{vY}{Average vertical speed [px/s].}
  \item{aX}{Average absolute horizontal acceleration [px/s^2].}
  \item{aY}{Average absolute vertical acceleration[px/s^2].}
  \item{flipsX}{Number of \emph{flips} (changing direction of move)
                on horizontal axis.}
  \item{flipsY}{Number of \emph{flips} (changing direction of move)
                on vertical axis.}
  \item{dX_sc, dY_sc, vX_sc, vY_sc, aX_sc, aY_sc, flipsX_sc, flipsY_sc}{
        \emph{Scrolling corrected} versions of the aforementioned indices.
        i.e. computed ruling out cursor moves over the survey screen that
        occurred because of the other actions than moving a pointing device
        (in particular: because of scrolling - either using mouse wheel,
        touchpad gestures or scrolling bar - using arrows to scroll the page
        or using TAB key to switch between survey form INPUT fields).}
}
Additionally, if \code{systemInfo} argument was provided there are included
columns:
\describe{
  \item{dX_rel, dY_rel, vX_rel, vY_rel, aX_rel, aY_rel, dX_screl, dY_screl,
        vX_screl, vY_screl, aX_screl, aY_screl}{
        \emph{Relative} versions of the aforementioned indices, i.e. divided
        by the width or height of a rectangle spanned by the most upper-left
        and the most bottom-right survey form INPUT element's (used to mark
        question answers) position on a given survey screen. These ones are
        better comparable between respondents with different browser window
        size.}
}
}
\description{
Function computes several indices describing the way respondent
moves a cursor on a survey screen.
}
\details{
Please note that values of average absolute acceleration (and only
these indices!) depends on whether \emph{stagnations} were previously
separated from \emph{mousemove} events or no - see
\code{\link{separate_stagnations}} for details.
}
\seealso{
\code{\link{separate_logdata_types}}
}
