% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cursor_positions.R
\name{compute_cursor_positions}
\alias{compute_cursor_positions}
\title{Computing cursor position at evenly spaced time points}
\usage{
compute_cursor_positions(
  actions,
  respId = any_of(c("id", "token", "respid")),
  screenId = all_of("screen"),
  timeSpan = 250
)
}
\arguments{
\item{actions}{A data frame containing data regarding \emph{actions} -
typically element \emph{actions} of a list returned by
\code{\link{separate_logdata_types}}.}

\item{respId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying respondent.}

\item{screenId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying survey screen.}

\item{timeSpan}{Space between time points in results in milliseconds.}
}
\value{
A data frame with columns:
\describe{
  \item{respId}{Column(s) defined by \code{respId}.}
  \item{screenId}{Column(s) defined by \code{screenId}.}
  \item{timeStampRel}{Time point (in milliseconds, counted from the time of
                      \emph{pageLoaded} event occurred.)}
  \item{pageX}{Last recorded cursor position on horizontal axis before
               a given \emph{timeStampRel}.}
  \item{pageY}{Last recorded cursor position on vertical axis before
               a given \emph{timeStampRel}.}
  \item{pageX_rel}{Last recorded \emph{relative} cursor position
                   (see \code{\link{compite_relative_positions}}) on
                   horizontal axis before a given \emph{timeStampRel}.
                   Included only if there was \emph{pageX_rel} column in the
                   input data.}
  \item{pageY_rel}{Last recorded \emph{relative} cursor position
                   (see \code{\link{compite_relative_positions}}) on
                   vertical axis before a given \emph{timeStampRel}.
                   Included only if there was \emph{pageY_rel} column in the
                   input data.}
}
}
\description{
Function compute the cursor positions at evenly spaced time
points to make the returned results useful, for example to draw heat maps.
}
\seealso{
\code{\link{compute_relative_positions}},
\code{\link{separate_logdata_types}}
}
