% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_editing.R
\name{compute_editing}
\alias{compute_editing}
\title{Computing number of (sub)question edits}
\usage{
compute_editing(
  actions,
  respId = any_of(c("id", "token", "respid")),
  screenId = "screen",
  entryId = c(),
  returnFormat = c("long", "wide")
)
}
\arguments{
\item{actions}{A data frame containing data regarding \emph{actions} -
typically element \emph{actions} of a list returned by
\code{\link{separate_logdata_types}}.}

\item{respId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}>  Variable(s)
identifying respondent.}

\item{screenId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying survey screen.}

\item{entryId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying survey screen \emph{entry} (compare \link{separate_returns}).
Set to \code{c()} to indicate that results should be returned for the whole survey
screens and not separately for each respondent's entry on a given survey
screen.}

\item{returnFormat}{String indicating whether results should be returned in
the \emph{long} (row is respondent-(sub)question) or in the \emph{wide}
(row is a respondent, there is a separate column for each (sub)question)
format. Can be abbreviated.}
}
\value{
If \code{returnFormat} is \emph{long} a data frame with columns:
\describe{
\item{respId}{Column(s) defined by \code{respId}.}
\item{screenId}{Column(s) defined by \code{screenId}.}
\item{entryId}{Column(s) defined by \code{entryId} (if any).}
\item{questionCode}{Code of a question.}
\item{subquestionCode}{Code of a subquestion.}
\item{edits}{Number of edits.}
}

If \code{returnFormat} is \emph{wide}, returned data frame will not contain columns
identifying screen, \emph{entry} (if applies), question and subquestion, but values
of these variables will be appended to column names (separated by "_"), with
each column reporting the number of edits of a given sub(question) for
a given screen (or screen-\emph{entry}).
}
\description{
Function computes number of edits (i.e. marking or changing an
answer) for each (sub)question made by each respondent. Be aware that
\strong{returned number of edits includes also first marking of an answer},
so correcting one's response is indicated by number of edits greater than 1.
}
\seealso{
\link{separate_logdata_types}
}
