% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hovering.R
\name{compute_hovering}
\alias{compute_hovering}
\title{Computing hovering indicators}
\usage{
compute_hovering(
  actions,
  respId = any_of(c("id", "token", "respid")),
  screenId = "screen",
  entryId = any_of("entry"),
  collapseAnswerCellAndControl = TRUE,
  showPB = TRUE
)
}
\arguments{
\item{actions}{A data frame containing data regarding \emph{actions} -
typically element \emph{actions} of a list returned by
\code{\link{separate_logdata_types}}.}

\item{respId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}>  Variable(s)
identifying respondent.}

\item{screenId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying survey screen.}

\item{entryId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying survey screen \emph{entry} (compare \link{separate_returns}).
Set to \code{c()} to indicate that results should be returned for the whole survey
screens and not separately for each respondent's entry on a given survey
screen.}

\item{collapseAnswerCellAndControl}{A logical value indicating
whether in case of table-format questions answer cells and form controls
placed inside these cells should be collapsed into one category in the
returned data or reported separately. \code{TRUE} by default.}

\item{showPB}{A logical value indicating whether to show a progress bar.
\code{TRUE} by default.}
}
\value{
A data frame with columns:
\describe{
\item{respId}{Column(s) defined by \code{respId}.}
\item{screenId}{Column(s) defined by \code{screenId}.}
\item{questionCode}{Question code (only if it is present in the input
data).}
\item{subquestionCode}{Subquestion code (only if it is present in the input
data).}
\item{answerCode}{Answer code (only if it is present in the input data).}
\item{elementType}{Type of the survey interface element.}
\item{hoverTime}{Total time spent by cursor over the element in seconds.}
}
}
\description{
Function sums up time spent by the cursor over specific elements
of the survey screens. If the input data frame contains columns
\emph{questionCode}, \emph{subquestionCode} and \emph{answerCode} (or at
least some of them) they will be taken into account while defining distinct
records of the returned results. Otherwise time will be summarized for each
respondent-screen over different values of the \emph{elementType} column.
}
\details{
Please note that you may change the way results are summarized by
selecting a specific subset of variables from a data frame you provide as
the \code{actions} argument. If you want to compute hovering simply over
different types of survey elements, exclude variables \emph{questionCode} and
\emph{subquestionCode} from the input data frame.
}
\seealso{
\link{separate_logdata_types}
}
