% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aat.R
\name{count_number_of_items}
\alias{count_number_of_items}
\title{Counting number of items on survey screens}
\usage{
count_number_of_items(x, screenId = "screen")
}
\arguments{
\item{x}{A data frame containing survey form INPUT elements' positions
(particularly the \emph{inputPositions} element of a list returned by
\link{separate_logdata_types}) or some other data frame containing survey screen
id, \emph{questionCode} and \emph{subquestionCode} variables.}

\item{screenId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying survey screen.}
}
\value{
A data frame with columns:
\describe{
\item{screenId}{Column(s) defined by \code{screenId}.}
\item{nQuestions}{Number of questions on a given survey screen.}
\item{nItes}{Number of \emph{items} on a given survey screen.}
}
}
\description{
This function may be used to prepare the \code{nItems} argument
to the \link{compute_aat} or \link{compute_aht} function call.

While counting number of \emph{items}:
\itemize{
\item{In questions containing subquestions each subquestion counts as
a single \emph{item}.}
\item{Each question that does not contain subquestions counts as a single
\emph{item}.}
}
}
\seealso{
\link{separate_logdata_types}
}
