% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logstreams_to_data_frame.R
\name{logstreams_to_data_frame}
\alias{logstreams_to_data_frame}
\title{Transforming a data frame with log-data streams to a data frame in
a long format}
\usage{
logstreams_to_data_frame(logData, respId, questionNamesTo, questionNamesPrefix)
}
\arguments{
\item{logData}{A data frame with log-data streams recorded by the
\emph{logDataLimeSurvey} applet stored in its columns.}

\item{respId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying respondent. (You may also list additional variables
here, that you want to keep joined with your log-data).}

\item{questionNamesTo}{A string - name of the variable that will identify
survey screen (page) in the results (i.e. one storing names of variables
storing log-data streams in the input data).}

\item{questionNamesPrefix}{A string - prefix of names of variables storing
log-data streams that should be deleted while turning these names into values
of the variable identifying survey screen.}
}
\value{
A data frame with columns:
\describe{
  \item{timeStamp}{Time stamp of an event.}
  \item{type}{Type of event.}
  \item{target.tagName}{Type of a HTML element that triggered an event (may
                        be empty for some types of events).}
  \item{target.id}{Id of a HTML element that triggered an event (may be empty
                   for some types of events).}
  \item{target.class}{CSS classes assigned to a HTML element that triggered
                      an event (may be empty for some types of events).}
  \item{which}{Information about which mouse button was pressed.}
  \item{metaKey}{Information regrading a meta key (ALT, CTRL) was pressed.}
  \item{pageX}{Location of an event on the webpage - horizontal axis.}
  \item{pageY}{Location of an event on the webpage - vertical axis.}
  \item{broken}{Flag indicating whether a record is broken.}
}
}
\description{
Function is called internally by
\code{\link{separate_logdata_types}} and is not exported (i.e. it is not
intended to be called by package users themselves). It pivots columns of
a data frame storing log-data streams to a \emph{long} format and then
separates log-data streams into different rows and columns. It performs
\strong{no} log-data consistency checks or additional transformations (like
assuring formats of specific columns) - this is done only afterwards by
another functions (see \code{\link{preprocess_actions}}).
}
\seealso{
\code{\link{separate_logdata_types}}
}
