% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_actions.R
\name{preprocess_actions}
\alias{preprocess_actions}
\title{Performing preprocessing of events}
\usage{
preprocess_actions(logData, respScreenIds, surveyStructure = NULL)
}
\arguments{
\item{logData}{data frame with log-data as returned by
\code{\link{logstreams_to_data_frame}}}

\item{respScreenIds}{<tidy-select> names of columns identifying
respondent-screens}

\item{surveyStructure}{data frame with mapping of \emph{SGQA identifiers}
onto question, subquestion and answer codes}
}
\value{
data frame with columns:
\itemize{
  \item{Columns defined by \code{respScreenIds}}
  \item{timeStamp,}
  \item{timeStampRel,}
  \item{type,}
  \item{target.tagName,}
  \item{target.id,}
  \item{target.class,}
  \item{which,}
  \item{metaKey,}
  \item{pageX,}
  \item{pageY,}
  \item{broken,}
  \item{moveX,}
  \item{moveY,}
  \item{duration,}
  \item{moveXScrollCorrected,}
  \item{moveYScrollCorrected,}
  \item{elementType.}
}
and if \code{surveyStructure} argument was provided with additional columns:
\itemize{
  \item{surveyId,}
  \item{questionId,}
  \item{questionFormat,}
  \item{SGQA,}
  \item{questionCode,}
  \item{subquestionCode,}
  \item{answerCode.}
}
}
\description{
Function is called internally by
\code{\link{separate_logdata_types}} and is not exported (i.e. it is not
intended to be called by package users themselves). It selects records
describing respondent \emph{actions} (i.e. actual events) and performs
a preprocessing that includes:
\itemize{
  \item{Converting column types storing numeric values to numeric type and
        substituting "undefined" values with \code{NAs}.}
  \item{Computing length of scrolling events (given information about
        \emph{page offsets} that is written in log-data streams;)}
  \item{Converting \emph{mousemove} events - that actually report a snapshot
        of a cursor position (that's how this works in JavaScript) - into
        events that actually describe moves}
  \item{Only if \code{surveyStructure} is provided: assigning question,
        subquestion (item) and response category codes to events connected
        with \emph{hovering} and provides a simple classification of what was
        \emph{hovered}.}
  \item{Assigning labels broadly describing type of an element on which
        a given event was triggered.}
}.
}
