% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_input_postions.R
\name{preprocess_input_positions}
\alias{preprocess_input_positions}
\title{Performing preprocessing of input positions}
\usage{
preprocess_input_positions(
  logData,
  respScreenIds,
  surveyStructure = NULL,
  imputeLastPageXY = median
)
}
\arguments{
\item{logData}{data frame with log-data as returned by
\code{\link{logstreams_to_data_frame}}}

\item{respScreenIds}{<tidy-select> names of columns identifying
respondent-screens}

\item{surveyStructure}{optionally a data frame with mapping of values
possibly occurring in the \code{target.id} column of a data frame provided by
the \code{logData} argument onto question, subquestion and answer codes as
returned by \code{\link{read_survey_structure}}}

\item{imputeLastPageXY}{function that will be used to impute position of
inputs in last subquestions/answers in array-format questions - see
\emph{Details} section below; it is \code{\link{median}} by default}
}
\value{
data frame with columns:
\itemize{
  \item{Columns defined by \code{respScreenIds}}
  \item{target.tagName,}
  \item{target.id,}
  \item{target.class,}
  \item{width,}
  \item{height,}
  \item{pageX,}
  \item{pageY.}
}
and if \code{surveyStructure} argument was provided with additional columns:
\itemize{
  \item{questionId,}
  \item{questionFormat,}
  \item{subquestionCode,}
  \item{answerCode.}
}
}
\description{
Function is called internally by
\code{\link{separate_logdata_types}} and is not exported (i.e. it is not
intended to be called by package users themselves). It selects records
describing input element's positions from a data frame storing different
types of log-data (returned by \code{\link{logstreams_to_data_frame}}) and
performs a preprocessing that includes:
\itemize{
  \item{Converting column types storing numeric values to numeric type.}
  \item{Transforming \code{pageX} and \code{pageY} columns so they describe
        position of the \strong{center} of the element instead of position
        of its upper-left corner.}
  \item{If  \code{surveyStructure} argument is provided, transforming
        \code{pageX} and \code{pageY} columns
         \strong{for inputs from array-type questions} so they describe
         actual position of an element on the page (there are some
         limitations - see \emph{Details} section below).}
}
Last two transformations make returned data more useful while creating
backgrounds on plots showing cursor moves.
}
\details{
Because of some aspects of how \emph{LimeSurvey's} responsive
layout works with respect to array type questions in these questions position
of INPUT elements do not describe actual position at which INPUT element were
visible for the respondent (i.e. center of an array cell; instead, these
positions describe the upper-left corner of a table cell in which a given
element was placed). \strong{Function can identify array type questions only
if \code{respScreenIds} argument is provided} and then it tries to correct
reported positions so they describe actual position of INPUT elements on the
page. However, there are some limitations:
\itemize{
  \item{Precise position can be derived for every answer/subquestion except
        the last ones by subtracting position reported for inputs associated
        with the next answer/subquestion from position of inputs associated
        with a given answer/subquestion (and dividing such a difference
        by 2).}
  \item{Position of the last answer/subquestion can not be derived and must
        be imputed. Size of the correction is computed using a function
        provided by \code{imputeLastPageXY} argument - \code{\link{median}}
        by default. This function is called on a vector of corrections that
        were already computed for the other answers/subquestions. With
        respect to the horizontal axis such imputation typically should be
        precise because \emph{LimeSurvey} rather creates table cells of the
        same width in array type questions. \strong{With respect to the
        vertical axis, it is nothing more than a lucky guess, however.}}
  \item{\strong{If there is only one answer or subquestion, position on
        a respective dimension can not be corrected (because of the lack of
        data) and it is therefore left unchanged.}}
  \item{At the moment \strong{no corrections are made for an array
        dual-scale questions}.}
}
}
