% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_system_info.R
\name{preprocess_system_info}
\alias{preprocess_system_info}
\title{Performing preprocessing of system information}
\usage{
preprocess_system_info(logData, inputPositions, respScreenIds)
}
\arguments{
\item{logData}{data frame with log-data as returned by
\code{\link{logstreams_to_data_frame}}}

\item{inputPositions}{data frame with data on INPUT element positions as
returned by \code{\link{preprocess_input_positions}}}

\item{respScreenIds}{<tidy-select> names of columns identifying
respondent-screens}
}
\value{
data frame with columns:
\itemize{
  \item{Columns defined by \code{respScreenIds}}
  \item{userAgent,}
  \item{language,}
  \item{browserWidth,}
  \item{browserHeight,}
  \item{screenWidth,}
  \item{screenHeight,}
  \item{inputsMinPageX,}
  \item{inputsMinPageY,}
  \item{inputsWidth.}
  \item{inputsHeight.}
}
}
\description{
Function is called internally by
\code{\link{separate_logdata_types}} and is not exported (i.e. it is not
intended to be called by package users themselves). It selects records
describing system information (i.e. respondent's browser and screen) and
performs a preprocessing that includes:
\itemize{
  \item{Renaming columns accordingly to the data they store.}
  \item{Converting column types storing numeric values to numeric type.}
  \item{Removing duplicates that occurred due to page being reloaded
        (typically because of some responses were not given to
        \emph{mandatory} questions or some criteria defined with respect to
        answers were not fulfilled).}
  \item{Putting data regarding browser and screen into one row (for each
        respondent-screen).}
  \item{Including information regarding position of the most upper-left
        INPUT element used to mark question answers and width and height of
        a rectangle spanned by the most upper-left and the most bottom-right
        INPUT element's - used to mark question answers -  position.
        This information is useful to perform standardization of cursor
        moves indices to correct for respondent-specific differences in
        survey layout.}
}
}
\seealso{
\code{\link{find_problems}}
}
