% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_logdata_types.R
\name{separate_logdata_types}
\alias{separate_logdata_types}
\title{Separating raw log-data into information on system, input position and
actions}
\usage{
separate_logdata_types(
  logData,
  surveyStructure = NULL,
  respId = any_of(c("id", "token", "respid")),
  questionNamesTo = "screen",
  questionNamesPrefix = "",
  inputsBoxCells = FALSE,
  separateReturns = TRUE,
  screenReturnThreshold = 1000
)
}
\arguments{
\item{logData}{A data frame with log-data streams recorded by the
\emph{logDataLimeSurvey} applet stored in its columns.}

\item{surveyStructure}{Optionally either a name (or vector of names) of a
\emph{LimeSurvey} survey structure file exported in the text (.txt) format or
a data frame with such a file already read, or list of data frames with such
files already read.}

\item{respId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying respondent. (You may also list additional variables
here, that you want to keep joined with your log-data).}

\item{questionNamesTo}{A string - name of the variable that will identify
survey screen (page) in the results (i.e. one storing names of variables
storing log-data streams in the input data).}

\item{questionNamesPrefix}{A string - prefix of names of variables storing
log-data streams that should be deleted while turning these names into values
of the variable identifying survey screen.}

\item{inputsBoxCells}{A logical value - whether \strong{in a case of table-format
questions} \code{inputsMinPageX}, \code{inputsMinPageY}, \code{inputsWidth} and
\code{inputsHeight} should be computed with respect to
boundaries (vertexes) of table cells storing the input elements rather than
with respect to input element positions. Applied only for table-format
question that are the only questions on a given page.}

\item{separateReturns}{A logical value indicating whether returns to the same
survey screen (different survey screen \emph{entries}) should be identified by an
additional id-column in the returned results.}

\item{screenReturnThreshold}{(Applicable only if \code{separateReturns != "no"})
The smallest number of milliseconds between the \emph{pageLoaded} event directly
following the \emph{submit} event that will be interpreted as indicating returning
to a given survey screen after visiting some other screen. If set to \code{Inf},
returns will be identified only by looking on the sequence of screen visits
(this have limitations if paradata is not collected on some survey screens)
\itemize{
\item see \emph{Details} in \link{separate_returns}.
}}
}
\value{
A list of three data frames with elements named \emph{systemInfo},
\emph{inputPositions} and  \emph{actions}. More information on how these
data frames are constructed you may find in the documentation regarding
functions listed below in the \emph{See also} section.
}
\description{
Function separates raw log-data streams into three separate data
frames containing:
\itemize{
\item{\emph{System information}: data on respondent's browser, OS,
screen and browser window resolution.}
\item{Information about the \emph{layout} of the survey screen.
Specifically, position of each INPUT element on the page. This
information will be useful to standardize cursor moves indicators
with respect to the differences in the survey page layout between
respondents and also to draw backgrounds to graphs presenting
cursor trajectories.}
\item{Information about respondent's \emph{actions}.}
}
Data frame provided as an input typically comes from reading a CSV file with
survey results that was previously exported from the LimeSurvey. It should
contain \strong{only} 1) responses to the questions (columns) that were used
to store log-data streams and 2) columns explicitly stated in the
\code{respId} argument (that need not be only ids, actually).
}
\seealso{
\link{read_survey_structure}, \link{logstreams_to_data_frame},
\link{separate_returns}, \link{preprocess_input_positions}, \link{preprocess_system_info},
\link{preprocess_actions}, \link{find_problems}, \link{separate_stagnations}
}
