% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_returns.R
\name{separate_returns}
\alias{separate_returns}
\title{Identifying and labelling returns to survey screens}
\usage{
separate_returns(logData, respId, screenId, screenReturnThreshold = 1000)
}
\arguments{
\item{logData}{A data frame with log-data as returned by
\code{\link{logstreams_to_data_frame}}.}

\item{respId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying respondent.}

\item{screenId}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> Variable(s)
identifying survey screen.}

\item{screenReturnThreshold}{The smallest number of milliseconds between the
\emph{pageLoaded} event directly following the \emph{submit} event that will be
interpreted as indicating returning to a given survey screen after visiting
some other screen. If set to \code{Inf}, returns will be identified only by
looking on the sequence of screen visits (this have limitations if paradata
is not collected on some survey screens) - see \emph{Details} section below.}
}
\description{
Function is called internally by \link{separate_logdata_types} and
is not exported (i.e. it is not intended to be called by package users
themselves). Function creates additional column \code{entry} that describes which
time a given screen was visited by a given respondent while a given event
occurred.
}
\details{
Identification of returns may be performed in two ways (each with its own
limitations):
\itemize{
\item{By looking at the sequences of screens as they appear in a log-stream
ordered by time stamps. However, these method works reliably only for
survey screens which are surrounded by another survey screens
collecting the log-data (and more generally, only if all the survey
screens collect log-data - if respondents can navigate to any survey
screen at any time). Otherwise it will be possible to visit
another screen without leaving a sign of if in the log-data.}
\item{By looking at the pairs of directly following each other \emph{submit}
and \emph{pageLoaded} events. The problem here is that if \emph{LimeSurvey}
encounters some incorrectly filled up questions while respondent is
trying to follow to the next survey screen, it reloads the page.
This results in recording the same sequence of the \emph{submit} event
directly followed by the \emph{pageLoaded} event as if respondent has
actually left the survey screen and returned. One may try to
distinguish between these situations by using a \strong{heuristic} -
assuming that a very short time (shorter than the one specified by
\code{screenReturnThreshold}) will indicate reloading the page and longer
times indicate actual leaving and returning to a give survey screen.}
}
By default the function uses both approaches described above. The second one
may be turned off by giving \code{screenReturnThreshold = Inf}.
}
