% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_stagnations.R
\name{separate_stagnations}
\alias{separate_stagnations}
\title{Separating periods of no move from mousemoves actions}
\usage{
separate_stagnations(actions, threshold)
}
\arguments{
\item{actions}{data frame containg data regarding \emph{actions} - typically
element \code{actions} of a list returned by
\code{\link{separate_logdata_types}}}

\item{threshold}{duration in milliseconds - \emph{mousemove} events lasting
longer will be separated}
}
\value{
data frame with the same columns as input but (probably) with some
additional rows of type \emph{mousemove}
}
\description{
Functions enables to separate (relatively) long-lasting
\emph{mousemove} \emph{actions} into two: one describing a period in which
a cursor not moved and another describing actual move.
}
\details{
JavaScript \emph{mousemove} event reports cursor position but
is triggered by moving a cursor. As a consequence respondent not moving
a cursor will be recorded in log-data returned by
\code{\link{separate_logdata_types}} as a (relatively) long-lasting (but
probably rather \emph{short-distanced}) \emph{mousemove}. This function
enables to separate such \emph{actions} into one covering a period of
stagnation (not moving a cursor) and another covering actual move. Sadly,
this can be done only by dividing \emph{action} on
\strong{arbitrarily chosen threshold} of duration - it is reasonable to set
it to the value of frequency of collecting \emph{mousemove} events that was
used in the LimeSurvey log-data collecting applet (typically 100 ms) or
a little bigger value.
}
