% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cursor_indices.R
\name{compute_cursor_indices}
\alias{compute_cursor_indices}
\title{Computing indicators describing cursor moves}
\usage{
compute_cursor_indices(
  actions,
  respScreenIds = all_of(intersect(names(actions), c("id", "token", "screen"))),
  systemInfo = NULL
)
}
\arguments{
\item{actions}{data frame containg data regarding \emph{actions} - typically
element \code{actions} of a list returned by
\code{\link{separate_logdata_types}}}

\item{respScreenIds}{<tidy-select> names of columns identifying
respondent-screens}

\item{systemInfo}{optionally data frame containg data regarding
\emph{system information} - typically element \code{systemInfo} of a list
returned by \code{\link{separate_logdata_types}}; it is required to compute
relative variants of indices}
}
\value{
data frame with columns:
\itemize{
  \item{Columns defined by \code{respScreenIds}}
  \item{\code{dX} - total distance travelled on horizontal axis,}
  \item{\code{dY} - total dostance travelled on vertical axis,}
  \item{\code{vX} - average horizontal speed,}
  \item{\code{vY} - average vertical speed,}
  \item{\code{aX} - average absolute horizontal acceleration,}
  \item{\code{aY} - average absolute vertical acceleration,}
  \item{\code{flipsX} - number of \emph{flips} (changing direction of move) on horizontal axis,}
  \item{\code{flipsY} - number of \emph{flips} (changing direction of move) on vertical axis,}
  \item{Aforementioned columns with suffix \emph{_sc} - analogous indiced
        computed using cursor moves corrected for scrolling (i.e. not taking
        into acount cursor moves on a survey screen that took place because
        of other actions than moving a cursor itself using a pointing device).}
}
and if \code{systemInfo} argument was provided with additional columns:
\itemize{
  \item{Aforementioned columns, with the exception of \code{flipsX} and
        \code{flipsY}, with suffix \emph{_rel} or \emph{_screl} - relative
        variants of the aforementioned indices, i.e. divided by the width or
        height of a rectangle spanned by the most upper-left and the most
        bottom-right INPUT element's - used to mark question answers -
        position on a given survey screen.}
}
}
\description{
Function computes few indices describing the way respondent
moves a cursor on a survey screen.
}
\seealso{
\code{\link{separate_logdata_types}}
\code{\link{preprocess_system_info}}
\code{\link{preprocess_input_positions}}
}
