% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_editing.R
\name{compute_editing}
\alias{compute_editing}
\title{Computing number of (sub)question edits}
\usage{
compute_editing(
  actions,
  respId = any_of(c("id", "token", "respid")),
  screenId = all_of("screen"),
  returnFormat = c("long", "wide")
)
}
\arguments{
\item{actions}{A data frame containing data regarding \emph{actions} -
typically element \emph{actions} of a list returned by
\code{\link{separate_logdata_types}}.}

\item{respId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying respondent.}

\item{screenId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying survey screen.}

\item{returnFormat}{String indicating whether results should be returned in
a \emph{long} (row is respondent-(sub)question) or in a \emph{wide}
(row is a respondent, there is a separate column for each (sub)question)
format. Can be abbreviated.}
}
\value{
If \code{returnFormat} is \emph{long} a data frame with columns:
\describe{
  \item{respId}{Column(s) defined by \code{respId}.}
  \item{screenId}{Column(s) defined by \code{screenId}.}
  \item{questionCode}{Code of a question.}
  \item{subquestionCode}{Code of a subquestion.}
  \item{edits}{Number of edits.}
}
If \code{returnFormat} is \emph{wide} a data frame with columns:
\describe{
  \item{respId}{Column(s) defined by \code{respId}.}
  \item{screenId}{Column(s) defined by \code{screenId}.}
  \item{edits_qCode_sqCode}{Columns named in a convention \emph{edits_codes},
                            where \emph{qCode} is question code and
                            \emph{sqCode} is subquestion code, storing number
                            of edits.}
}
}
\description{
Function computes number of edits (i.e. marking or changing an
answer) for each (sub)question made by each respondent. Be aware that
\strong{returned number of edits includes also first marking of an answer},
so correcting one's response is indicated by number of edits greater than 1.
}
\seealso{
\code{\link{separate_logdata_types}}
}
