% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_relative_positions.R
\name{compute_relative_positions}
\alias{compute_relative_positions}
\title{Computing relative positions of mousemove events}
\usage{
compute_relative_positions(
  actions,
  systemInfo,
  respId = any_of(c("id", "token", "respid")),
  screenId = all_of("screen")
)
}
\arguments{
\item{actions}{A data frame containing data regarding \emph{actions} -
typically element \emph{actions} of a list returned by
\code{\link{separate_logdata_types}}.}

\item{systemInfo}{A data frame containing data regarding
\emph{system information} - typically element \code{systemInfo} of a list
returned by \code{\link{separate_logdata_types}}.}

\item{respId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying respondent.}

\item{screenId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying survey screen.}
}
\value{
Input data frame with additional columns:
\describe{
  \item{pageX_rel}{\emph{Relative} position of an event on horizontal axis.}
  \item{pageY_rel}{\emph{Relative} position of an event on vertical axis.}
  \item{moveX_rel}{\emph{Relative} distance of a move on horizontal axis.}
  \item{moveY_rel}{\emph{Relative} distance of a move on vertical axis.}
}
Additionally, if \code{moveXScrollCorrected_} and \emph{moveYScrollCorrected}
columns are available in the input data, there are included columns:
\describe{
  \item{moveXScrollCorrected_rel}{\emph{Relative scroll corrected} distance
                                  of a move on horizontal axis.}
  \item{moveYScrollCorrected_rel}{\emph{Relative scroll corrected} distance
                                  of a move on vertical axis.}
}
}
\description{
Function computes position of \emph{mousemove} events that are
expressed relatively to the positions of the most upper-left and bottom-right
INPUT elements (used to mark responses) of the survey form. Such recalculated
positions are better comparable between respondents (with possible
differences in their hardware and software configurations) and are
particularly useful for drawing heat-maps or cursor traces.
}
\seealso{
\code{\link{compute_cursor_indices}},
\code{\link{compute_cursor_positions}},
\code{\link{separate_logdata_types}}
}
