% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey_structure.R
\name{read_survey_structure}
\alias{read_survey_structure}
\title{Reading in text files storing survey structure}
\usage{
read_survey_structure(x)
}
\arguments{
\item{x}{character vector with file names \strong{or} a data frame with
already read survey structure file}
}
\value{
data frame with columns:
\itemize{
  \item{target.id,}
  \item{surveyId,}
  \item{questionId,}
  \item{questionFormat,}
  \item{SGQA,}
  \item{questionCode,}
  \item{subquestionCode,}
  \item{answerCode.}
}
}
\description{
Function is called internally by
\code{\link{separate_logdata_types}} and is not exported (i.e. it is not
intended to be called by package users themselves). It reads in text
(tab-delimited) files storing survey structure previously exported from the
\emph{LimeSurvey} web-survey platform (if provided with file names) or passes
down a provided data frame.
}
