% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{remove_problems}
\alias{remove_problems}
\title{Removing problems identified while separating log-data}
\usage{
remove_problems(
  x,
  systemInfo = NULL,
  respId = any_of(c("id", "token", "respid")),
  screenId = all_of("screen")
)
}
\arguments{
\item{x}{Either a list returned by \code{\link{separate_logdata_types}} or
element \emph{actions} of such a list. In the latter case also argument
\code{systemInfo} must be provided.}

\item{systemInfo}{A data frame containing data regarding
\emph{system information} - typically element \code{systemInfo} of a list
returned by \code{\link{separate_logdata_types}}.}

\item{respId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying respondent.}

\item{screenId}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Variable(s) identifying survey screen.}
}
\value{
Input data (either a data frame or a list of three data frames)
possibly with some rows removed.
}
\description{
Function enables to remove from log-data object prepared by
\code{\link{separate_logdata_types}} those records that were identified
(by this function) to be somehow problematic. Function works interactively
presenting user description of the problem and number of records affected
and enables to choose whether given records should be removed or kept.
}
\details{
Every column which name starts with \emph{problems} in the
\emph{systemInfo} element of \code{x} or in \code{systemInfo} argument will
be treated as marking problems in the data.
}
\seealso{
\code{\link{separate_logdata_types}},
\code{\link{separate_stagnations}}
}
