% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_logdata_types.R
\name{separate_logdata_types}
\alias{separate_logdata_types}
\title{Separating raw log-data into information on system input position and
actions}
\usage{
separate_logdata_types(
  logData,
  surveyStructure = NULL,
  respId = all_of(intersect(names(logData), c("id", "token"))),
  questionNamesTo = "screen"
)
}
\arguments{
\item{logData}{data frame with log-data streams recorded by the
\emph{logDataLimeSurvey} applet stored in its columns}

\item{surveyStructure}{optionally either a name (or vector of names) of a
\emph{LimeSurvey} survey structure file exported in the text (.txt) format or
a data frame with such a file already read, or list of data frames with such
files already read}

\item{respId}{optionally (default value covers typical structure of
survey results exported from the \emph{LimeSurvey}) <tidy-select> name(s)
of the column(s) containing respondent's id (and perhaps other variables you
want to keep joined with your log-data)}

\item{questionNamesTo}{optionally a string - name of a variable that will
identify (describe) survey screen (page) in the results}
}
\value{
List of three data frames with elements named \emph{systemInfo},
\emph{inputPositions} and  \emph{actions}. Some more information on how these
data frames are constructed you may find in the documentation regarding
functions listed below in the \emph{See also} section.
}
\description{
Function separates raw log-data streams into three separate data
frames containing:
\itemize{
  \item{\emph{System information}: data on respondent's browser, OS,
        screen and browser window resolution.}
  \item{Information about the \emph{layout} of the web page: specifically,
        position of each INPUT element on the page. This information will be
        useful to standardize cursor moves indicators with respect to the
        differences in the survey page layout between respondents.}
  \item{Information about respondent's \emph{actions}.}
}
Data frame provided as an input typically comes from reading a CSV file with
survey results that was previously exported from the LimeSurvey. It should
contain \strong{only} 1) responses to the questions (columns) that were used
to store log-data streams and 2) columns explicitly stated in the
\code{respId} argument (that need not be only ids, actually).
}
\seealso{
\code{\link{read_survey_structure}}
\code{\link{logstreams_to_data_frame}}
\code{\link{preprocess_input_positions}} \code{\link{preprocess_system_info}}
\code{\link{find_problems}} \code{\link{preprocess_actions}}
}
